+++ src/controller_device_state.c
@@ -815,6 +815,8 @@ device_config_read_cache(clixon_handle h,
     char  *db;
     cxobj *xt = NULL;
     cxobj *xroot;
+    cxobj *xerr = NULL;
+    int    ret;
 
     if (devname == NULL || config_type == NULL){
         clixon_err(OE_UNIX, EINVAL, "devname or config_type is NULL");
@@ -826,14 +828,23 @@ device_config_read_cache(clixon_handle h,
     }
     cprintf(cb, "device-%s-%s", devname, config_type);
     db = cbuf_get(cb);
-    if (xmldb_get_cache(h, db, &xt, NULL) < 0)
+    if ((ret = xmldb_get_cache(h, db, &xt, &xerr)) < 0)
         goto done;
+    if (ret == 0){
+        if ((*cberr = cbuf_new()) == NULL){
+            clixon_err(OE_UNIX, errno, "cbuf_new");
+            goto done;
+        }
+        if (netconf_err2cb(h, xerr, *cberr) < 0)
+            goto done;
+        goto failed;
+    }
     if ((xroot = xpath_first(xt, NULL, "devices/device/config")) == NULL){
         if ((*cberr = cbuf_new()) == NULL){
             clixon_err(OE_UNIX, errno, "cbuf_new");
             goto done;
         }
-        cprintf(*cberr, "No such device tree");
+        cprintf(*cberr, "Datastore %s does not contain device tree: devices/device/config", db);
         goto failed;
     }
     if (xdatap){
@@ -841,6 +852,8 @@ device_config_read_cache(clixon_handle h,
     }
     retval = 1;
  done:
+    if (xerr)
+        xml_free(xerr);
     if (cb)
         cbuf_free(cb);
     return retval;